refactor(py): rename googleai_lyria.py to lyria.py for consistency
